IPointGenerator.cs and CircularCloudLayouter.cs were refactored. CirclePointGenerator.cs and EllipsePointGenerator.cs were added
