+++ TagCloud/PointGenerator/CirclePointGenerator.cs
@@ -1,0 +1,12 @@
+﻿using System.Drawing;
+
+namespace TagCloud.PointGenerator
+{
+    public class CirclePointGenerator : ArchimedeanSpiral
+    {
+        public CirclePointGenerator(Point centralPoint) 
+            : base(centralPoint, 1, 1)
+        {
+        }
+    }
+}
