+++ TagCloud/TagCloud.csproj
@@ -85,15 +85,17 @@
     <Compile Include="AppConfig\IAppConfigProvider.cs" />
     <Compile Include="App\ConsoleApp.cs" />
     <Compile Include="App\IApp.cs" />
-    <Compile Include="CloudLayouter\ArchimedeanSpiral.cs" />
+    <Compile Include="PointGenerator\ArchimedeanSpiral.cs" />
+    <Compile Include="PointGenerator\CirclePointGenerator.cs" />
     <Compile Include="CloudLayouter\CircularCloudLayouter.cs" />
+    <Compile Include="PointGenerator\EllipsePointGenerator.cs" />
     <Compile Include="CommandLineParsing\Options.cs" />
     <Compile Include="ContainerConfig.cs" />
     <Compile Include="FileReader\DocxFileReader.cs" />
     <Compile Include="ImageProcessing\ICloudImageGenerator.cs" />
     <Compile Include="ImageProcessing\ImageSettings.cs" />
     <Compile Include="CloudLayouter\Vector.cs" />
-    <Compile Include="CloudLayouter\IPointGenerator.cs" />
+    <Compile Include="PointGenerator\IPointGenerator.cs" />
     <Compile Include="WordColoring\BlackColoring.cs" />
     <Compile Include="WordColoring\GradientColoring.cs" />
     <Compile Include="WordColoring\IWordColoring.cs" />
