+++ TagCloudUnitTests/ArchimedeanSpiralTests.cs
@@ -2,7 +2,7 @@
 using System.Drawing;
 using FluentAssertions;
 using NUnit.Framework;
-using TagCloud.CloudLayouter;
+using TagCloud.PointGenerator;
 
 namespace TagCloudUnitTests
 {
